Swapped FilterAggregator for FilterInterpreter
The FilterInterpreter and the FilterTokenizer shares the new
filteroperators.json data file which declares operators and their names.
The filteroperators.json is not truthful to users regarding what operators
exist and will be fixed soon.

The actual object-set iteration that previously was in the FilterAggregator
has been moved to the user-space i.e. the user of the Interpreter has to
run the .filter (array) method on their object set.

Signed-off-by: Tobias Sjöndin <tsjondin@op5.com>