Heavy modifications to make interpreting bearable
The FilterParserVisitor no longer consumes the tokens and let them remain
tokens so that the Interpreter can act on token types instead of their represented
value (how do you differ a list of values from an expression represented as a list?).

Had my produced expression tree been a binary expression tree with associativity I
would not have done this, and rather flagged values that caused issues with what they
were, but since that was not the case I felt it was appropriate to move in this direction
instead to make the interpreting easier.

Signed-off-by: Tobias Sjöndin <tsjondin@op5.com>